requirements.txt
# Borgo-AI Dependencies
# Optimized for RTX 3060 Ti (8GB VRAM) + 16GB RAM

# Core CLI
typer==0.21.1
rich==13.7.1

# LLM & Embeddings
requests==2.31.0

# Vector Store (CPU version - GPU optional)
faiss-cpu==1.9.0
numpy==1.24.4

# Web Scraping
beautifulsoup4==4.12.3
lxml==4.9.4

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
chardet==5.2.0

# Image Processing
Pillow==10.2.0

# Optional: GPU acceleration for FAISS (uncomment if desired)
# faiss-gpu>=1.7.4

# Optional: For better web parsing
# html2text>=2020.1.16
# trafilatura>=1.6.0
